Changes to dossan aalin
for some reason some cells of the text, as it appears on github, are repeated and are in red and green. Some of cells in white are not appearing on the website and I'm not sure if this update will change those issues. Will forward a word doc showing how it should look.