Ard-choonceil
This piece was forwarded to me by George Broderick in a document that contained some other 'Fockle ayns Dy Cleash' columns from 1974/75, and according to the order this is the first part of the column that was published on March 1st 1975. That issue of Manx Star is missing from the Museum microfiche.